# DistrictEnterpriseStatisticalData.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DistrictEnterpriseStatisticalData.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/supplier": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "summary": "Получение всех поставщиков",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supplier"
        ],
        "summary": "Создание поставщика",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Supplier"
        ],
        "summary": "Обновление информации о поставщике",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplier/{id}": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "summary": "Получение поставщика по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Supplier"
        ],
        "summary": "Удаление поставщика",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplier/supply-between-dates": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "summary": "Получение информации о всех поставщиках, поставивших сырье за заданных период, упорядоченных по названию",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/supplier/max-provided-suppliers": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "summary": "Получение информации о поставщиках, поставивших максимальное количество товара за указанный период",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/supply": {
      "get": {
        "tags": [
          "Supply"
        ],
        "summary": "Получение всех поставок",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supply"
        ],
        "summary": "Создание поставки",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Supply"
        ],
        "summary": "Обновление информации о поставке",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supply/{id}": {
      "get": {
        "tags": [
          "Supply"
        ],
        "summary": "Получение поставки по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Supply"
        ],
        "summary": "Удаление поставки",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EnterpriseDto": {
        "required": [
          "address",
          "employees_number",
          "enterprise_type",
          "form_of_ownership",
          "name",
          "phone_number",
          "total_area"
        ],
        "type": "object",
        "properties": {
          "registration_number": {
            "type": "integer",
            "description": "Регистрационный номер",
            "format": "int32"
          },
          "enterprise_type": {
            "$ref": "#/components/schemas/EnterpriseTypeDto"
          },
          "name": {
            "type": "string",
            "description": "Наименование",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "form_of_ownership": {
            "$ref": "#/components/schemas/FormOfOwnershipDto"
          },
          "employees_number": {
            "type": "integer",
            "description": "Количество работающих",
            "format": "int32"
          },
          "total_area": {
            "type": "integer",
            "description": "Общая площадь",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO предприятия"
      },
...
      "EnterpriseTypeDto": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Тип предприятия",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO типа предприятия"
      },
...
      "FormOfOwnershipDto": {
        "required": [
          "form"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "form": {
            "type": "string",
            "description": "Форма собственности",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO формы собственности"
      },
...
      "SupplierCreateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя поставщика",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания поставщика"
      },
      "SupplierDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор поставщика",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Имя поставщика",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO поставщика"
      },
...
      "SupplyCreateDto": {
        "required": [
          "date",
          "enterprise_registration_number",
          "quantity",
          "supplier_id"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "description": "Количество поставляемого сырья",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "description": "Дата поставки",
            "format": "date"
          },
          "enterprise_registration_number": {
            "type": "integer",
            "description": "Предприятие, которому осуществлялась поставка",
            "format": "int32"
          },
          "supplier_id": {
            "type": "integer",
            "description": "Поставщик",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания поставки"
      },
      "SupplyDto": {
        "required": [
          "date",
          "enterprise_registration_number",
          "quantity",
          "supplier_id"
        ],
        "type": "object",
        "properties": {
          "supply_id": {
            "type": "integer",
            "description": "Идентификатор поставки",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "description": "Количество поставляемого сырья",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "description": "Дата поставки",
            "format": "date"
          },
          "enterprise_registration_number": {
            "$ref": "#/components/schemas/EnterpriseDto"
          },
          "supplier_id": {
            "$ref": "#/components/schemas/SupplierDto"
          }
        },
        "additionalProperties": false,
        "description": "DTO поставки"
      }
    }
  }
}